Actually prevents SD3 in gen panel
