+++ src/components/Generation/Error/ErrorBoundary.tsx
@@ -1,8 +1,6 @@
-import { Title, Text, Button } from '@mantine/core';
 import type { ErrorInfo, ReactNode } from 'react';
 import React, { Component } from 'react';
-import { TwCard } from '~/components/TwCard/TwCard';
-import { generationFormStore, generationStore } from '~/store/generation.store';
+import { ResetGenerationPanel } from '~/components/Generation/Error/ResetGenerationPanel';
 
 interface Props {
   children: ReactNode;
@@ -26,8 +24,6 @@ class GenerationErrorBoundary extends Component<Props, State> {
 
   resetErrorBoundary() {
     this.setState({ hasError: false });
-    generationStore.clearData();
-    generationFormStore.reset();
   }
 
   componentDidCatch(error: Error, errorInfo: ErrorInfo) {
@@ -47,33 +43,11 @@ class GenerationErrorBoundary extends Component<Props, State> {
   render() {
     if (!this.state.hasError) return this.props.children;
     return (
-      <div className="flex size-full flex-col items-center justify-center p-2">
-        <div className="mb-5 flex flex-col items-center">
-          <div className="overflow-hidden rounded-xl shadow">
-            {/* eslint-disable-next-line @next/next/no-img-element */}
-            <img
-              src="/images/media/whoops.jpg"
-              alt="something went wrong"
-              className="w-full max-w-[200px]"
-            />
-          </div>
-          <br />
-          <Title order={3}>{`Something went wrong :(`}</Title>
-          <Button
-            onClick={() => {
-              const keys = Object.keys(localStorage).filter((key) =>
-                key.startsWith('generation-form')
-              );
-              for (const key of keys) {
-                localStorage.removeItem(key);
-              }
-              this.resetErrorBoundary();
-            }}
-          >
-            Reset Generator State
-          </Button>
-        </div>
-      </div>
+      <ResetGenerationPanel
+        onResetClick={() => {
+          this.resetErrorBoundary();
+        }}
+      />
     );
   }
 }
