+++ src/components/ImageGeneration/GenerationForm/GenerationForm2.tsx
@@ -135,6 +135,7 @@ import {
 } from '~/shared/orchestrator/hidream.config';
 import classes from './GenerationForm2.module.scss';
 import type { GenerationResource } from '~/server/services/generation/generation.service';
+import { ResetGenerationPanel } from '~/components/Generation/Error/ResetGenerationPanel';
 
 let total = 0;
 const tips = {
@@ -512,7 +513,14 @@ export function GenerationFormContent() {
           const disableSafetyTolerance = !isFluxKontext;
 
           const resourceTypes = getBaseModelResourceTypes(baseModel);
-          if (!resourceTypes) return <></>;
+          if (!resourceTypes)
+            return (
+              <ResetGenerationPanel
+                onResetClick={() => {
+                  form.reset();
+                }}
+              />
+            );
 
           return (
             <>
